Fix namespace for archive tests
